+++ packages/ci-context/src/lib/utils/local.ts
@@ -1,20 +1,5 @@
-import { getExecOutput } from '@nx-tools/core';
 import { RepoMetadata, RunnerContext } from '../interfaces';
-
-export const getRef = async () => {
-  const output = await getExecOutput('git symbolic-ref HEAD');
-  return output.stdout.trim();
-};
-
-export const getCommitUserEmail = async () => {
-  const output = await getExecOutput('git log -1 --pretty=format:%ae');
-  return output.stdout.trim();
-};
-
-export const getSha = async () => {
-  const output = await getExecOutput('git rev-parse HEAD');
-  return output.stdout.trim();
-};
+import { getCommitUserEmail, getRef, getSha } from './local-helpers';
 
 export async function context(): Promise<RunnerContext> {
   return {